+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file. The format 
 
 ## [Unreleased]
 
+## [10.1.0] - 2026-02-26
+
 ### Changed
 - Make `Mesh.x`, `Mesh.y` and `Mesh.z` writable.
 - Use the pseudo-inverse instead of the inverse of the Hessian to calculate `dx` in `ConstitutiveMaterial.optimize()`.